fix sticky property of nav bar
This was due to the html tag using 100% of the screen. Which was
required for keeping the footer at the bottom of the page. Instead, I
now have it so that the body (everything) must take up at least 100% of
the screen (using vh). Then the content (which is the wrapper for the
main block) will absorb any free space.