Add project pages (#24)
* feat: add projects home page

Just a page that can send you to pty projects or our projects (not sure
on that second name yet). This is to save us from a drop down within a
dropdown menu

* feat: add projects page to navbar

* refactor: change events layout to a general name

To be repurposed as something that can be used by project pages as well.

* fix: broken link

Links are put into lowercase (for directories at least)

* feat: add pages for projects

* refactor: get all blog pages to use the same template

DRY and all that. God damn was this a pain to get right. I don't think
you can define blocks in partials which was causing the main problems.

* fix: enforce consistent naming convention

* feat: add some example projects

* refactor: fix typo

* refactor: add left margin to featured box

* refactor: give outfit font to button

* style: fix indentation

* refactor: remove repeated flex-start fields

* feat: add archetype for creating new blogs

* fix: force image centering on the page (not the div)

* fix: broken links

* fix: get all profile pictures to be same size

* style: fix indentation

* refactor: change to projects

* feat: add ability to cite the cover image of a blog article