refactor: move people to json file (#35)
* feat: add json file covering people

* refactor: Use a json file for people instead of pages

refactor: remove old individual people pages

* feat: add a new way of changing the personnel

Hopefully this is easy enough to work through. I've done an initial test
to see how easy it is to understand. More people will be required to
improve it in the future.

* refactor: remove unecessary file

No longer using name parse in script

* refactor: remove email parsing

This is no longer required and is also not what we want. Alumni won't
have emails for @exeter.ac.uk anymore.

* feat: add new information for siyi