feat: overhaul of header (#40)
The nav bar now works by expanding down into the header, instead of into
the body. This also has had a nice side benefit of making the logo
bigger and therefore a little easier to see what it is.

It is a little more complex however, to get the menu to work flawlessly
(hopefully) it requires some javascript (to ensure the visibility works
as intended).