feat: Add PhD projects page (#43)
* refactor: remove useless page

* feat: add PhD projects link

refactor: Use a grid instead of subsequent divs

* feat: add background for phd projects link

* feat: add phd projects page

* feat: add example blog

* fix: back button returns to previous page

Previously was hard coded to return to events page

* refactor: implement max image size into blog entries

* refactor: remove centering of blog images

In general this looks better than a centered image (centered images only
look good when they are large and wide)

* refactor: control image size of featured articles

* feat: add back button to project blog selection pages

This should make navigation a little easier. This back button will not
show on the event page due to a new parameter.