feat: Add dark mode (#77)
* feat: add (non-functional) dark mode button

This is just the button, it doesn't do anything yet. Just switches
between a sun and a moon

* feat: add initial functional dark mode

currently only works for a single paragraph. But it does the thing

* feat: make publicaitons work with dark mode

Fuck me this does look awful though

* feat: remember the theme between pages

Problem: There is a small delay of turning to dark mode when switching
pages

* feat: add text colouring to people pages

* refactor: rename to more appropriate file name

* fix: stop screen flickering when using dark mode

* fix: page no longer reloads on button click

Previously there was an anchor element that was linking to the same
page. This causes the page to technically be reloaded when it was being
clicked. This resulted in the page moving back to the top whenever the
theme was toggled.