refactor: Use some more sass features (#79)
* refactor: rename colours to be more general

* fix: remove deprecated feature

sass 2.0 can no longer create global variables on the fly. I thought
this would be breaking, but in reality adding this line shuts the
transpiler up. The variable exists so it can now work with it.

* refactor: set theme-map back to null after setting it

* refactor: overhaul the font system to use variables

* refactor: move helper scss to partials folder

* refactor: move all buttons to a partial

Back buttons need a second mixin as they need certain margin and
position properties to look good.

* refactor: use nesting properties of sass