feat: Use default profile when picture is missing (#118)
* feat: use default profile if profile doesn't exist

* feat: add proper error message

* feat: add example usage (Rhian Swarbrick)

* refactor: move default profile picture to correct directory

Yes it is a svg, but it makes more sense to go next to the other people

* refactor: consistency with blog pages

Find images via name field only

* refactor: profile picture is not required

* fix: update scripts to work with default profile picture

* feat: add better error message if python is not found

* feat: add colours to error messages

* fix: correct index

Removing "image" means "alumni" is now the 2nd element of the
dictionary. As such it was being skipped over.

* feat: update logging

* feat: use realpaths (looks nicer in logging)

* feat: make info blue coloured

* feat: only display picture logging if adding people

* feat: ensure that correct input is given for change type

* style: add section headers

* feat: ensure name is given

* style: improve legibility

Heredocs are great, but goddamn the inability to indent them makes for
some real ugly lookin code ay?

* docs: update to reflect changes

* feat: add exit if picture is too large.

I can see people just moving past this if they don't care. But it really
is critical, so I'm enforcing the change.

* docs: reintroduce advice for getting smaller images

* docs(typo): spell recommend correctly